feat: refactor cylinder intersection .c file for norminette
