+++ Makefile
@@ -6,7 +6,7 @@
 #    By: chenlee <chenlee@student.42kl.edu.my>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/08/14 21:28:59 by chenlee           #+#    #+#              #
-#    Updated: 2023/11/03 22:15:58 by chenlee          ###   ########.fr        #
+#    Updated: 2023/11/04 11:57:19 by chenlee          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,31 +27,22 @@ OBJS_DIR	=	objects/
 OBJS		=	$(addprefix $(OBJS_DIR), $(notdir $(SRC:.c=.o)))
 
 # source files here #
-# SRC			=	error.c				\
-# 				initialization.c	\
-# 				get_next_line.c		\
-# 				parse.c				\
-# 				parse_helpers.c		\
-# 				add_shapes.c		\
-# 				add_single.c		\
-# 				general_utils.c		\
-# 				parse_utils.c		\
-# 				geom_trans.c		\
-# 				vector_algebra.c	\
-# 				vector_arithmetic.c	\
-# 				get_cy_dist.c		\
-# 				set_controls.c		\
-# 				free.c
-
-SRC			=	error.c				\
-				initialization.c	\
-				get_next_line.c		\
-				geom_trans.c		\
-				vector_algebra.c	\
-				vector_arithmetic.c	\
-				cy_intersection.c	\
-				get_sp_dist.c		\
-				set_controls.c		\
+SRC			=	error.c					\
+				initialization.c		\
+				get_next_line.c			\
+				parse.c					\
+				parse_helpers.c			\
+				add_shapes.c			\
+				add_single.c			\
+				general_utils.c			\
+				parse_utils.c			\
+				geom_trans.c			\
+				vector_algebra.c		\
+				vector_arithmetic.c		\
+				vector_utils.c			\
+				cy_intersection.c		\
+				cy_intersection_utils.c	\
+				set_controls.c			\
 				free.c
 
 # source directory here #
@@ -62,7 +53,7 @@ SRC_DIR		=	$(LIBX)			\
 				src/mem_free	\
 				src/ray_tracing	\
 				src/utils		\
-				# src/parsing		\
+				src/parsing		\
 
 vpath %.c $(SRC_DIR)
