+++ includes/mathematics.h
@@ -46,7 +46,6 @@ double	dot_prod(t_coord a, t_coord b);
 */
 t_coord	cross_prod(t_coord a, t_coord b);
 t_coord	normalize(t_coord a);
-int	approx(double value, double limit);
 
 /**************************** vector_arithmetic.c ****************************/
 
@@ -73,5 +72,28 @@ t_coord vect_add(t_coord a, t_coord b);
  * @return Function returns a resulting vector of the vectors parsed.
 */
 t_coord vect_subt(t_coord a, t_coord b);
+
+/******************************* vector_utils.c *******************************/
+
+/**
+ * @brief Function checks for each coordinate of a 3d vector, to determine if
+ * the vector is a zero vector ( <0, 0, 0> ). An incrementing integer is used to
+ * keep track of every coordinate if its value is not empty. If the integer
+ * remained zero, that means neither of the coordinate has a value, hence is a
+ * zero vector.
+ * @param vector The vector parsed for verification
+ * @return Returns true if vector is zero vector; otherwise, returns false
+*/
+int	is_zero_vector(t_coord vector);
+
+/**
+ * @brief Function checks if a double value is approximately the same as the
+ * limit, especially when direct equal comparison for doubles is not accurate.
+ * @param value THe double for comparison.
+ * @param limit The limit to be compared with.
+ * @return Returns true if the value is approximately the limit; otherwise,
+ * returns false.
+*/
+int	approx(double value, double limit);
 
 #endif
