+++ src/math/vector_utils.c
@@ -1,0 +1,43 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   vector_utils.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: chenlee <chenlee@student.42kl.edu.my>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/04 11:43:00 by chenlee           #+#    #+#             */
+/*   Updated: 2023/11/04 11:47:54 by chenlee          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minirt.h"
+
+int	approx(double value, double limit)
+{
+	double	difference;
+
+	if (value <= -0.0001)
+		value *= -1;
+	difference = value - limit;
+	if (difference > -0.0001 && difference < 0.0001)
+		return (1);
+	else
+		return (0);
+}
+
+int	is_zero_vector(t_coord vector)
+{
+	int	coord_not_zero;
+
+	coord_not_zero = 0;
+	if (vector.x > 0.00001 || vector.x < -0.00001)
+		coord_not_zero++;
+	if (vector.y > 0.00001 || vector.y < -0.00001)
+		coord_not_zero++;
+	if (vector.z > 0.00001 || vector.y < -0.00001)
+		coord_not_zero++;
+	if (coord_not_zero == 0)
+		return (1);
+	else
+		return (0);
+}
